requirements.txt
amqp==5.3.1
asgiref==3.10.0
azure-core==1.35.1
azure-storage-blob==12.26.0
billiard==4.2.2
boto3==1.40.50
botocore==1.40.50
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.2
Django==5.2.7
django-filter==25.2
django-storages==1.14.6
djangorestframework==3.16.1
gunicorn==23.0.0
idna==3.11
isodate==0.7.2
jmespath==1.0.1
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.9.0

